Fixed "Project type is not supported by this installation" error
IN VWD Express 2010, this error manifests due to having 3+ GUIDs in the
ProjecTypeGuids XML node.

Specific Error Messages:
On Open: One or more projects in the solution were not loaded correctly.
Please see Output Window for details.

After open, in Output: The project type is not supported by this
installation.